Fix census menu on elections (votings)
