+++ decidim-elections/lib/decidim/votings/census_admin_engine.rb
@@ -1,6 +1,6 @@
 # frozen_string_literal: true
 
-require "decidim/votings/menu"
+require "decidim/votings/census_menu"
 
 module Decidim
   module Votings
@@ -11,7 +11,7 @@ module Decidim
       paths["lib/tasks"] = nil
 
       initializer "decidim_votings_census.admin_voting_menu" do
-        Decidim::Votings::Menu.register_admin_voting_menu!
+        Decidim::Votings::CensusMenu.register_admin_voting_menu!
       end
     end
   end
